+++ tests/python/contrib/test_cmsisnn/test_binary_ops.py
@@ -36,12 +36,10 @@ from utils import (
     assert_partitioned_function,
     assert_no_external_function,
 )
+from tvm.testing.aot import generate_ref_data, AOTTestModel, compile_and_run
 from tvm.micro.testing.aot_test_utils import (
-    AOTTestModel,
     AOT_CORSTONE300_RUNNER,
     AOT_USMP_CORSTONE300_RUNNER,
-    generate_ref_data,
-    compile_and_run,
 )
