+++ tests/python/contrib/test_cmsisnn/test_conv2d.py
@@ -23,15 +23,9 @@ import tvm
 from tvm import relay
 from tvm.relay.op.contrib import cmsisnn
 
+from tvm.testing.aot import generate_ref_data, AOTTestModel, compile_and_run
 
-from tvm.micro.testing.aot_test_utils import (
-    AOTTestModel,
-    AOT_CORSTONE300_RUNNER,
-    AOT_USMP_CORSTONE300_RUNNER,
-    AOT_DEFAULT_RUNNER,
-    generate_ref_data,
-    compile_and_run,
-)
+from tvm.micro.testing.aot_test_utils import AOT_USMP_CORSTONE300_RUNNER
 from utils import (
     skip_if_no_reference_system,
     make_module,
