+++ tests/python/contrib/test_cmsisnn/test_invalid_graphs.py
@@ -22,12 +22,9 @@ import pytest
 import tvm
 from tvm import relay
 
-
+from tvm.testing.aot import AOTTestModel, compile_and_run, generate_ref_data
 from tvm.micro.testing.aot_test_utils import (
-    AOTTestModel,
     AOT_USMP_CORSTONE300_RUNNER,
-    generate_ref_data,
-    compile_and_run,
 )
 from utils import (
     skip_if_no_reference_system,
