+++ tests/python/relay/strategy/arm_cpu/test_conv2d_nchw.py
@@ -20,11 +20,9 @@ import pytest
 import tvm
 import tvm.testing
 from tvm import relay
+from tvm.testing.aot import AOTTestModel, compile_and_run, generate_ref_data
 from tvm.micro.testing.aot_test_utils import (
-    AOTTestModel,
     AOT_CORSTONE300_RUNNER,
-    generate_ref_data,
-    compile_and_run,
 )
